Refactor: Project Singular page now attempts to get an existing payload before attempting a require // Refactor: projects array in projects store is now compiled in the getProjects action for all routes
